claims/services/radar_sync.py
"""
RADAR Synchronization Service

This service handles all RADAR API integration and data synchronization.
"""
from typing import Dict, List, Any, Optional
from datetime import datetime
from django.conf import settings
from django.utils import timezone
import logging

logger = logging.getLogger(__name__)


class RADARSyncService:
    """Service class for RADAR system synchronization"""

    def __init__(self, api_url: Optional[str] = None, api_key: Optional[str] = None):
        """
        Initialize RADAR sync service

        Args:
            api_url: RADAR API base URL (defaults to settings.RADAR_API_URL)
            api_key: RADAR API key (defaults to settings.RADAR_API_KEY)
        """
        self.api_url = api_url or getattr(settings, 'RADAR_API_URL', '')
        self.api_key = api_key or getattr(settings, 'RADAR_API_KEY', '')
        self.retry_attempts = getattr(settings, 'RADAR_SYNC_RETRY_ATTEMPTS', 3)
        self.retry_delay = getattr(settings, 'RADAR_SYNC_RETRY_DELAY', 5)

    def sync_voyages(self, since: Optional[datetime] = None) -> Dict[str, Any]:
        """
        Synchronize voyages from RADAR

        Args:
            since: Only sync voyages updated since this timestamp

        Returns:
            dict: Sync results with counts of created/updated records
        """
        logger.info(f"Starting voyage sync from RADAR (since: {since})")

        try:
            # TODO: Implement actual RADAR API integration
            # This is a placeholder implementation

            from claims.models import Voyage

            # Placeholder: Mock API response
            # voyages_data = self._fetch_from_radar('/voyages', params={'since': since})

            created_count = 0
            updated_count = 0
            error_count = 0

            # TODO: Process each voyage from RADAR
            # for voyage_data in voyages_data:
            #     try:
            #         voyage, created = Voyage.objects.update_or_create(
            #             radar_voyage_id=voyage_data['id'],
            #             defaults={
            #                 'voyage_number': voyage_data['voyage_number'],
            #                 'vessel_name': voyage_data['vessel_name'],
            #                 # ... map other fields
            #                 'radar_data': voyage_data,
            #                 'last_radar_sync': timezone.now(),
            #             }
            #         )
            #         if created:
            #             created_count += 1
            #         else:
            #             updated_count += 1
            #     except Exception as e:
            #         logger.error(f"Error syncing voyage {voyage_data.get('id')}: {e}")
            #         error_count += 1

            return {
                'success': True,
                'created': created_count,
                'updated': updated_count,
                'errors': error_count,
                'timestamp': timezone.now().isoformat(),
            }

        except Exception as e:
            logger.error(f"RADAR voyage sync failed: {e}")
            return {
                'success': False,
                'error': str(e),
                'timestamp': timezone.now().isoformat(),
            }

    def sync_claims(self, since: Optional[datetime] = None) -> Dict[str, Any]:
        """
        Synchronize claims from RADAR

        Args:
            since: Only sync claims updated since this timestamp

        Returns:
            dict: Sync results with counts of created/updated records
        """
        logger.info(f"Starting claim sync from RADAR (since: {since})")

        try:
            # TODO: Implement actual RADAR API integration
            from claims.models import Claim

            created_count = 0
            updated_count = 0
            error_count = 0

            # TODO: Process each claim from RADAR

            return {
                'success': True,
                'created': created_count,
                'updated': updated_count,
                'errors': error_count,
                'timestamp': timezone.now().isoformat(),
            }

        except Exception as e:
            logger.error(f"RADAR claim sync failed: {e}")
            return {
                'success': False,
                'error': str(e),
                'timestamp': timezone.now().isoformat(),
            }

    def sync_port_activities(self, since: Optional[datetime] = None) -> Dict[str, Any]:
        """
        Synchronize port activities from RADAR

        Args:
            since: Only sync activities updated since this timestamp

        Returns:
            dict: Sync results with counts of created/updated records
        """
        logger.info(f"Starting port activity sync from RADAR (since: {since})")

        try:
            # TODO: Implement actual RADAR API integration
            from port_activities.models import PortActivity

            created_count = 0
            updated_count = 0
            error_count = 0

            # TODO: Process each activity from RADAR

            return {
                'success': True,
                'created': created_count,
                'updated': updated_count,
                'errors': error_count,
                'timestamp': timezone.now().isoformat(),
            }

        except Exception as e:
            logger.error(f"RADAR activity sync failed: {e}")
            return {
                'success': False,
                'error': str(e),
                'timestamp': timezone.now().isoformat(),
            }

    def sync_all(self, since: Optional[datetime] = None) -> Dict[str, Any]:
        """
        Synchronize all data from RADAR

        Args:
            since: Only sync data updated since this timestamp

        Returns:
            dict: Combined sync results
        """
        logger.info("Starting full RADAR synchronization")

        voyages_result = self.sync_voyages(since)
        claims_result = self.sync_claims(since)
        activities_result = self.sync_port_activities(since)

        return {
            'success': all([
                voyages_result.get('success'),
                claims_result.get('success'),
                activities_result.get('success'),
            ]),
            'voyages': voyages_result,
            'claims': claims_result,
            'port_activities': activities_result,
            'timestamp': timezone.now().isoformat(),
        }

    def _fetch_from_radar(self, endpoint: str, params: Optional[Dict] = None) -> List[Dict]:
        """
        Fetch data from RADAR API

        Args:
            endpoint: API endpoint (e.g., '/voyages')
            params: Optional query parameters

        Returns:
            list: API response data
        """
        # TODO: Implement actual HTTP request to RADAR API
        # This is a placeholder

        import requests
        from claims_system.utils import retry_on_failure

        @retry_on_failure(
            max_attempts=self.retry_attempts,
            delay=self.retry_delay,
            exponential_backoff=True
        )
        def make_request():
            url = f"{self.api_url}{endpoint}"
            headers = {
                'Authorization': f'Bearer {self.api_key}',
                'Content-Type': 'application/json',
            }
            response = requests.get(url, params=params, headers=headers, timeout=30)
            response.raise_for_status()
            return response.json()

        # return make_request()
        return []  # Placeholder

    def push_to_radar(self, data_type: str, data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Push data to RADAR system

        Args:
            data_type: Type of data (e.g., 'claim', 'voyage')
            data: Data to push

        Returns:
            dict: Push result
        """
        logger.info(f"Pushing {data_type} to RADAR")

        try:
            # TODO: Implement actual push to RADAR API
            return {
                'success': True,
                'timestamp': timezone.now().isoformat(),
            }

        except Exception as e:
            logger.error(f"Failed to push {data_type} to RADAR: {e}")
            return {
                'success': False,
                'error': str(e),
                'timestamp': timezone.now().isoformat(),
            }
