requirements.txt
asgiref==3.11.0
Django==5.2.9
django-storages==1.14.5
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3

# Environment variables
python-decouple==3.8

# Security & Rate Limiting
django-ratelimit==4.1.0

# Logging
python-json-logger==3.2.1

# Testing & Coverage
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.10

# Code Quality
flake8==7.1.1
black==24.10.0
mypy==1.13.0
django-stubs==5.1.1

# Background Tasks (Celery)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.8.0
django-celery-results==2.5.1

# Caching
django-redis==5.4.0

# Security Scanning
bandit==1.8.0
safety==3.2.14

# Pre-commit hooks
pre-commit==4.0.1

# REST API
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-filter==24.3
drf-spectacular==0.27.2
